% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rows.R
\name{as.data.table}
\alias{as.data.table}
\title{Convert an object to a data table, with a log}
\usage{
as.data.table(x, ...)
}
\arguments{
\item{x}{The object to convert.}

\item{...}{All other arguments of \code{\link[data.table:as.data.table]{data.table::as.data.table()}}.}
}
\value{
The \code{data.table} that \code{\link[data.table:as.data.table]{data.table::as.data.table()}} returns.
}
\description{
Convert an object to a data table, with a log
}
\examples{
data.table::as.data.table(head(mtcars, 3), keep.rownames = "car")
}
