% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group.R
\name{grouping_sets}
\alias{grouping_sets}
\alias{rollup}
\alias{cube}
\alias{groupingsets}
\title{Aggregate over several grouping sets, with a log}
\usage{
rollup(x, ...)

cube(x, ...)

groupingsets(x, ...)
}
\arguments{
\item{x}{The data table to aggregate.}

\item{...}{All other arguments, as in the corresponding \code{data.table}
function.}
}
\value{
Whatever the \code{data.table} function returns.
}
\description{
\code{\link[data.table:rollup]{data.table::rollup()}}, \code{\link[data.table:cube]{data.table::cube()}} and
\code{\link[data.table:groupingsets]{data.table::groupingsets()}} run one aggregation per grouping set and stack
the results. Each reports the size of the table it produced next to the size
of the table it started from, in the wording an aggregating \code{j} uses.
}
\examples{
dt <- data.table::data.table(g = c("a", "a", "b"), h = c("x", "y", "x"),
                             v = 1:3)
data.table::rollup(dt, j = sum(v), by = c("g", "h"))
data.table::cube(dt, j = sum(v), by = c("g", "h"))
}
