% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rows.R
\name{head_tail}
\alias{head_tail}
\alias{head.data.table}
\alias{tail.data.table}
\title{First or last rows of a data table, with a log}
\usage{
\method{head}{data.table}(x, ...)

\method{tail}{data.table}(x, ...)
}
\arguments{
\item{x}{The data table.}

\item{...}{All other arguments of \code{\link[utils:head]{utils::head()}} and \code{\link[utils:tail]{utils::tail()}}, i.e.
\code{n}.}
}
\value{
The same rows \code{data.table} would return.
}
\description{
First or last rows of a data table, with a log
}
\examples{
head(data.table::as.data.table(mtcars), 3)
}
