% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape.R
\name{reshape}
\alias{reshape}
\alias{melt}
\alias{melt.data.table}
\alias{dcast}
\alias{dcast.data.table}
\title{Reshape a data table, with a log}
\usage{
melt(data, ...)

\method{melt}{data.table}(data, ...)

dcast(data, ...)

\method{dcast}{data.table}(data, ...)
}
\arguments{
\item{data}{The table to reshape.}

\item{...}{All other arguments of \code{\link[data.table:melt.data.table]{data.table::melt.data.table()}} and
\code{\link[data.table:dcast.data.table]{data.table::dcast.data.table()}}.}
}
\value{
The reshaped data table.
}
\description{
Reports which columns were reorganized into which, and how the dimensions
of the table changed, in the style of \code{tidylog}'s \code{pivot_longer()} and
\code{pivot_wider()} messages.
}
\examples{
dt <- data.table::data.table(id = 1:2, a = 3:4, b = 5:6)
long <- data.table::melt(dt, id.vars = "id")
data.table::dcast(long, id ~ variable)
}
