% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{fread}
\alias{fread}
\title{Read a file into a data table, with a log}
\usage{
fread(...)
}
\arguments{
\item{...}{All arguments of [data.table::fread()].}
}
\value{
The data table that [data.table::fread()] returns.
}
\description{
Read a file into a data table, with a log
}
\examples{
fread(text = "a,b\n1,2\n3,4")
}
