% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{fwrite}
\alias{fwrite}
\title{Write a data table to a file, with a log}
\usage{
fwrite(x, ...)
}
\arguments{
\item{x}{The table to write.}

\item{...}{All other arguments of [data.table::fwrite()].}
}
\value{
`NULL`, invisibly, as [data.table::fwrite()] returns it.
}
\description{
Write a data table to a file, with a log
}
\examples{
dt <- data.table::data.table(a = 1:2, b = 3:4)
data.table::fwrite(dt, tempfile())
}
