% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rows.R
\name{head_tail}
\alias{head_tail}
\alias{head.data.table}
\alias{tail.data.table}
\title{First or last rows of a data table, with a log}
\usage{
\method{head}{data.table}(x, ...)

\method{tail}{data.table}(x, ...)
}
\arguments{
\item{x}{The data table.}

\item{...}{All other arguments of [utils::head()] and [utils::tail()], i.e.
`n`.}
}
\value{
The same rows `data.table` would return.
}
\description{
First or last rows of a data table, with a log
}
\examples{
head(data.table::as.data.table(mtcars), 3)
}
