% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape.R
\name{reshape}
\alias{reshape}
\alias{melt}
\alias{melt.data.table}
\alias{dcast}
\alias{dcast.data.table}
\title{Reshape a data table, with a log}
\usage{
melt(data, ...)

melt.data.table(data, ...)

dcast(data, ...)

dcast.data.table(data, ...)
}
\arguments{
\item{data}{The table to reshape.}

\item{...}{All other arguments of [data.table::melt.data.table()] and
[data.table::dcast.data.table()].}
}
\value{
The reshaped data table.
}
\description{
Reports which columns were reorganized into which, and how the dimensions
of the table changed, in the style of `tidylog`'s `pivot_longer()` and
`pivot_wider()` messages.
}
\examples{
dt <- data.table::data.table(id = 1:2, a = 3:4, b = 5:6)
long <- data.table::melt(dt, id.vars = "id")
data.table::dcast(long, id ~ variable)
}
