% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rows.R
\name{rows}
\alias{rows}
\alias{unique.data.table}
\alias{duplicated.data.table}
\alias{na.omit.data.table}
\alias{rbindlist}
\alias{funion}
\alias{fintersect}
\alias{fsetdiff}
\title{Row operations with a log}
\usage{
\method{unique}{data.table}(x, ...)

\method{duplicated}{data.table}(x, ...)

\method{na.omit}{data.table}(object, ...)

rbindlist(l, ...)

funion(x, y, ...)

fintersect(x, y, ...)

fsetdiff(x, y, ...)
}
\arguments{
\item{x, y, object, l}{The inputs, as in the corresponding `data.table`
function.}

\item{...}{All other arguments, passed on unchanged.}
}
\value{
Whatever the `data.table` function returns.
}
\description{
These functions behave exactly like their `data.table` counterparts and
report how many rows they removed, kept or combined.
}
\examples{
dt <- data.table::data.table(a = c(1, 1, 2), b = c(NA, 2, 3))
unique(dt, by = "a")
stats::na.omit(dt)
}
