% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_functions.R
\name{set_functions}
\alias{set_functions}
\alias{setnames}
\alias{setcolorder}
\alias{setkey}
\alias{setkeyv}
\alias{setorder}
\alias{setorderv}
\alias{setindex}
\alias{setindexv}
\alias{set}
\alias{setDT}
\alias{setDF}
\alias{setattr}
\title{Modify a data table by reference, with a log}
\usage{
setnames(x, ...)

setcolorder(x, ...)

setkey(x, ...)

setkeyv(x, ...)

setorder(x, ...)

setorderv(x, ...)

setindex(x, ...)

setindexv(x, ...)

set(x, ...)

setDT(x, ...)

setDF(x, ...)

setattr(x, ...)
}
\arguments{
\item{x}{The data table (or, for [setDT()], the object to convert).}

\item{...}{All other arguments, passed on unchanged.}
}
\value{
Whatever the corresponding `data.table` function returns.
}
\description{
These functions are the `data.table` `set*()` functions. They still change
their input by reference and return exactly what `data.table` returns; they
only report what they changed.
}
\examples{
dt <- data.table::data.table(a = 3:1, b = 1:3)
data.table::setnames(dt, "a", "alpha")
data.table::setkey(dt, alpha)
}
